# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}/follow:
    post:
      tags: [social]
      operationId: toggleFollow
      summary: Follow or unfollow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/{userId}/profile:
    get:
      tags: [social]
      operationId: getUserProfile
      summary: Get a user's profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    patch:
      tags: [social]
      operationId: updateUserProfile
      summary: Update own profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileBody'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /me/profile:
    get:
      tags: [social]
      operationId: getMyProfile
      summary: Get current user's profile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags: [social]
      operationId: updateMyProfile
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileBody'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    FollowResponse:
      type: object
      required: [following, followerCount]
      properties:
        following:
          type: boolean
        followerCount:
          type: integer
...
    UserProfile:
      type: object
      required: [userId, followerCount, followingCount, postCount, studioHours, isVerified, createdAt]
      properties:
        userId:
          type: string
        handle:
          type: [string, 'null']
        displayName:
          type: [string, 'null']
        bio:
          type: [string, 'null']
        medium:
          type: [string, 'null']
        location:
          type: [string, 'null']
        website:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
        bannerUrl:
          type: [string, 'null']
        isVerified:
          type: boolean
        isFollowing:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
        postCount:
          type: integer
        studioHours:
          type: integer
        kilnStatus:
          type: [string, 'null']
        generation:
          type: [integer, 'null']
        createdAt:
          type: string
    UpdateProfileBody:
      type: object
      properties:
        handle:
          type: string
        displayName:
          type: string
        bio:
          type: string
        medium:
          type: string
        location:
          type: string
        website:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        kilnStatus:
          type: string
...